refactoring: split some parts at different files
