+++ src/symbols.js
@@ -1,0 +1,3 @@
+export const $registeredMacros = Symbol('registeredMacros');
+export const $macroState = Symbol('macroState');
+export const $processedByMacro = Symbol('processedByMacro');
